execute inline
