+++ developer-console/src/main/java/in/ramanujan/developer/console/Console.java
@@ -7,10 +7,16 @@ import java.util.List;
 
 public class Console {
     public static void main(String[] args) throws IOException {
+        // If no arguments, do nothing
         if(args.length == 0) {
-            System.out.println("No arguments given");
             return;
         }
+        // If one argument is given, treat it as a path and use execute_inline
+        if(args.length == 1) {
+            OperationType.execute_inline.getImplementation().execute(Arrays.asList(args));
+            return;
+        }
+        // If more than one argument, use the first as operation type
         OperationType operationType = OperationType.getOperation(args[0]);
         if(operationType == null) {
             List<String> possibleOperators = new ArrayList<>();
